refactor: 총 금액에 대한 부분 타입 변경
- 범위 초과를 생각하여 타입 변경

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -2,16 +2,16 @@ package lotto.domain;
 
 public class LottoPrize {
 
-    private final int amount;
+    private final long amount;
     private final Currency currency;
 
-    public LottoPrize(int amount,
+    public LottoPrize(long amount,
                       Currency currency) {
         this.amount = amount;
         this.currency = currency;
     }
 
-    public int calculateAmount(Integer value) {
+    public long calculateAmount(long value) {
         return this.amount * value;
     }
